requirements.txt
python-snap7==1.2
